refactor: split notebook support helpers (#2132)
